+++ kauldron/konfig/__init__.py
@@ -41,9 +41,11 @@ cfg = konfig.resolve(cfg)
 # pylint: disable=g-importing-member
 from kauldron.konfig.configdict_base import ConfigDict
 from kauldron.konfig.configdict_base import register_aliases
+from kauldron.konfig.configdict_base import register_default_values
 from kauldron.konfig.configdict_proxy import resolve
 from kauldron.konfig.fake_import_utils import imports
 from kauldron.konfig.fake_import_utils import mock_modules
+from kauldron.konfig.fake_import_utils import set_lazy_imported_modules
 from kauldron.konfig.ref_utils import ref_copy
 from kauldron.konfig.ref_utils import ref_fn
 from kauldron.konfig.ref_utils import WithRef