+++ CHANGELOG.md
@@ -8,6 +8,8 @@ Changelog follow the https://keepachangelog.com/ standard (at least the headers)
 
 ## [Unreleased]
 
+* Changed: `konfig.imports()` are not lazy by default anymore (config don't
+  need to be resolved in `with ecolab.adhoc()` anymore!)
 * Added:
   * `kd.optim`: Optimizer / optax utils
   * `kd.eval`: Eval moved to their separate namespace